+++ R/vntrs-package.R
@@ -1,5 +1,6 @@
 #' @aliases vntrs-package
 #' @useDynLib vntrs, .registration = TRUE
 #' @importFrom Rcpp evalCpp
+#' @inherit vntrs examples
 #' @keywords internal
 "_PACKAGE"